Refactor presampling/downsampling
