+++ modyn/selector/internal/selector_strategies/__init__.py
@@ -9,7 +9,7 @@ from .freshness_sampling_strategy import FreshnessSamplingStrategy  # noqa: F401
 from .gradnorm_downsampling_strategy import GradNormDownsamplingStrategy  # noqa: F401
 from .loss_downsampling_strategy import LossDownsamplingStrategy  # noqa: F401
 from .new_data_strategy import NewDataStrategy  # noqa: F401
-from .random_downsampling_strategy import RandomDownsamplingStrategy  # noqa: F401
+from .random_presampling_strategy import RandomPresamplingStrategy  # noqa: F401
 
 files = os.listdir(os.path.dirname(__file__))
 files.remove("__init__.py")
