--- modyn/selector/internal/selector_strategies/random_downsampling_strategy.py
@@ -1,7 +1,0 @@
-from modyn.selector.internal.selector_strategies.abstract_downsample_strategy import AbstractDownsampleStrategy
-
-
-class RandomDownsamplingStrategy(AbstractDownsampleStrategy):
-    def __init__(self, config: dict, modyn_config: dict, pipeline_id: int, maximum_keys_in_memory: int):
-        super().__init__(config, modyn_config, pipeline_id, maximum_keys_in_memory)
-        self._requires_remote_computation = False
