+++ modyn/tests/selector/internal/selector_strategies/test_actual_samplers.py
@@ -5,6 +5,7 @@ import tempfile
 
 import pytest
 from modyn.metadata_database.metadata_database_connection import MetadataDatabaseConnection
+from modyn.selector.internal.selector_strategies import GradNormDownsamplingStrategy, RandomPresamplingStrategy
 from modyn.selector.internal.selector_strategies.loss_downsampling_strategy import LossDownsamplingStrategy
 
 database_path = pathlib.Path(os.path.abspath(__file__)).parent / "test_storage.db"
@@ -36,7 +37,7 @@ def setup_and_teardown():
     shutil.rmtree(TMP_DIR)
 
 
-def test_init():
+def test_init_loss():
     # Test init works
     strat = LossDownsamplingStrategy(
         {"limit": -1, "reset_after_trigger": False, "presampling_ratio": 80, "downsampled_batch_size": 10},
@@ -50,9 +51,23 @@ def test_init():
     assert isinstance(strat.get_downsampling_strategy(), str)
 
 
-def test_command():
+def test_init_gradnorm():
     # Test init works
-    strat = LossDownsamplingStrategy(
+    strat = GradNormDownsamplingStrategy(
+        {"limit": -1, "reset_after_trigger": False, "presampling_ratio": 80, "downsampled_batch_size": 10},
+        get_minimal_modyn_config(),
+        42,
+        1000,
+    )
+
+    assert strat.downsampled_batch_size == 10
+    assert strat._pipeline_id == 42
+    assert isinstance(strat.get_downsampling_strategy(), str)
+
+
+def test_command_gradnorm():
+    # Test init works
+    strat = GradNormDownsamplingStrategy(
         {"limit": -1, "reset_after_trigger": False, "presampling_ratio": 80, "downsampled_batch_size": 10},
         get_minimal_modyn_config(),
         42,
@@ -62,6 +77,34 @@ def test_command():
     name = strat.get_downsampling_strategy()
     params = strat.get_downsampling_params()
     assert isinstance(name, str)
-    assert name == "RemoteLossDownsampling"
+    assert name == "RemoteGradNormDownsampling"
     assert "downsampled_batch_size" in params
     assert params["downsampled_batch_size"] == 10
+
+
+def test_init_random():
+    # Test init works
+
+    with pytest.raises(ValueError):
+        RandomPresamplingStrategy(
+            {"limit": -1, "reset_after_trigger": False, "presampling_ratio": 80, "downsampled_batch_size": 10},
+            get_minimal_modyn_config(),
+            42,
+            1000,
+        )
+
+    strat = RandomPresamplingStrategy(
+        {"limit": -1, "reset_after_trigger": False, "presampling_ratio": 80},
+        get_minimal_modyn_config(),
+        42,
+        1000,
+    )
+
+    assert strat._pipeline_id == 42
+    assert not hasattr(strat, "downsampled_batch_size")
+
+    with pytest.raises(NotImplementedError):
+        strat.get_downsampling_strategy()
+
+    with pytest.raises(NotImplementedError):
+        strat.get_downsampling_params()
